+++ config/config.exs
@@ -1,22 +1,28 @@
 import Config
 
-config :prima_auth0_ex,
-  auth0_base_url: "https://tenant.eu.auth0.com"
+# Default client, for backwards compatibility
+config :prima_auth0_ex, :clients,
+  default_client: [
+    auth0_base_url: "https://your-auth0-provider.com",
+    client_id: "",
+    client_secret: "",
+    cache_namespace: "my-service",
+    token_check_interval: :timer.seconds(1),
+    signature_check_interval: :timer.seconds(1)
+  ]
 
-config :prima_auth0_ex, :client,
-  client_id: "",
-  client_secret: "",
-  cache_enabled: true,
-  cache_namespace: "my-service",
-  cache_encryption_key: "uhOrqKvUi9gHnmwr60P2E1hiCSD2dtXK1i6dqkU4RTA=",
-  redis_connection_uri: "redis://redis:6379",
-  redis_ssl_enabled: false,
-  redis_ssl_allow_wildcard_certificates: false
+config :prima_auth0_ex, :redis,
+  enabled: true,
+  encryption_key: "uhOrqKvUi9gHnmwr60P2E1hiCSD2dtXK1i6dqkU4RTA=",
+  connection_uri: "redis://redis:6379",
+  ssl_enabled: false,
+  ssl_allow_wildcard_certificates: false
 
 config :prima_auth0_ex, :server,
+  auth0_base_url: "https://your-auth0-provider.com",
   ignore_signature: false,
   audience: "some-audience",
-  issuer: "https://tenant.eu.auth0.com/",
+  issuer: "https://your-auth0-tenant.com",
   first_jwks_fetch_sync: true
 
 config :logger, :console, metadata: :all