+++ lib/prima_auth0_ex.ex
@@ -8,17 +8,17 @@ defmodule PrimaAuth0Ex do
   @doc """
   Obtain a token for the given audience.
   """
-  @spec token_for(String.t()) :: {:ok, String.t()} | {:error, any()}
-  def token_for(audience) do
-    TokenProvider.token_for(TokenProvider, audience)
+  @spec token_for(String.t(), atom()) :: {:ok, String.t()} | {:error, any()}
+  def token_for(audience, client \\ :default_client) do
+    TokenProvider.token_for(client, audience)
   end
 
   @doc """
   Force the refresh of the token for a given audience, invalidating both the local and the shared cache.
   """
-  @spec refresh_token_for(String.t()) :: {:ok, String.t()} | {:error, any()}
-  def refresh_token_for(audience) do
-    TokenProvider.refresh_token_for(TokenProvider, audience)
+  @spec refresh_token_for(String.t(), atom()) :: {:ok, String.t()} | {:error, any()}
+  def refresh_token_for(audience, client \\ :default_client) do
+    TokenProvider.refresh_token_for(client, audience)
   end
 
   @doc """
@@ -30,7 +30,8 @@ defmodule PrimaAuth0Ex do
   When `ignore_signature` is `true`, only checks the validity of claims of the token and not its signature.
   This option should never be enabled in production-like environments, as it allows anyone to forge valid tokens.
   """
-  @spec verify_and_validate(String.t(), String.t()) :: {:ok, Joken.claims()} | {:error, atom | Keyword.t()}
+  @spec verify_and_validate(String.t(), String.t()) ::
+          {:ok, Joken.claims()} | {:error, atom | Keyword.t()}
   @spec verify_and_validate(String.t(), String.t(), list(String.t())) ::
           {:ok, Joken.claims()} | {:error, atom | Keyword.t()}
   @spec verify_and_validate(String.t(), String.t(), list(String.t()), boolean()) ::