+++ lib/prima_auth0_ex/token_provider/cached_token_service.ex
@@ -8,21 +8,27 @@ defmodule PrimaAuth0Ex.TokenProvider.CachedTokenService do
   instances will retrieve the new token from the new cache and will not have to generate a new
   one from the authorization provider.
   """
-  alias PrimaAuth0Ex.TokenProvider.{Auth0AuthorizationService, EncryptedRedisTokenCache, TokenService}
+  alias PrimaAuth0Ex.Config
+
+  alias PrimaAuth0Ex.TokenProvider.{
+    Auth0AuthorizationService,
+    EncryptedRedisTokenCache,
+    TokenService
+  }
 
   @behaviour TokenService
 
   @impl TokenService
   def retrieve_token(credentials, audience) do
-    audience
-    |> token_cache().get_token_for()
+    credentials.client
+    |> token_cache().get_token_for(audience)
     |> refresh_token_on_cache_miss(credentials, audience)
   end
 
   @impl TokenService
   def refresh_token(credentials, audience, current_token, false = _force_cache_bust) do
-    audience
-    |> token_cache().get_token_for()
+    credentials.client
+    |> token_cache().get_token_for(audience)
     |> refresh_token_unless_it_changed(current_token, credentials, audience)
   end
 
@@ -30,9 +36,17 @@ defmodule PrimaAuth0Ex.TokenProvider.CachedTokenService do
     do_refresh_token(credentials, audience)
   end
 
-  defp refresh_token_on_cache_miss({:error, _}, credentials, audience), do: do_refresh_token(credentials, audience)
-  defp refresh_token_on_cache_miss({:ok, nil}, credentials, audience), do: do_refresh_token(credentials, audience)
-  defp refresh_token_on_cache_miss({:ok, cached_token}, _credentials, _audience), do: {:ok, cached_token}
+  defp refresh_token_on_cache_miss(result, credentials, audience)
+
+  defp refresh_token_on_cache_miss({:error, _}, credentials, audience),
+    do: do_refresh_token(credentials, audience)
+
+  defp refresh_token_on_cache_miss({:ok, nil}, credentials, audience),
+    do: do_refresh_token(credentials, audience)
+
+  defp refresh_token_on_cache_miss({:ok, cached_token}, _, _), do: {:ok, cached_token}
+
+  defp refresh_token_unless_it_changed(result, token, credentials, audience)
 
   defp refresh_token_unless_it_changed({:error, _}, _, credentials, audience) do
     do_refresh_token(credentials, audience)
@@ -42,7 +56,8 @@ defmodule PrimaAuth0Ex.TokenProvider.CachedTokenService do
     do_refresh_token(credentials, audience)
   end
 
-  defp refresh_token_unless_it_changed({:ok, cached_token}, current_token, _, _) when cached_token != current_token do
+  defp refresh_token_unless_it_changed({:ok, cached_token}, current_token, _, _)
+       when cached_token != current_token do
     {:ok, cached_token}
   end
 
@@ -53,7 +68,7 @@ defmodule PrimaAuth0Ex.TokenProvider.CachedTokenService do
   defp do_refresh_token(credentials, audience) do
     case authorization_service().retrieve_token(credentials, audience) do
       {:ok, token} ->
-        token_cache().set_token_for(audience, token)
+        token_cache().set_token_for(credentials.client, audience, token)
         {:ok, token}
 
       {:error, description} ->
@@ -62,7 +77,7 @@ defmodule PrimaAuth0Ex.TokenProvider.CachedTokenService do
   end
 
   defp authorization_service,
-    do: Application.get_env(:prima_auth0_ex, :authorization_service, Auth0AuthorizationService)
+    do: Config.authorization_service(Auth0AuthorizationService)
 
-  defp token_cache, do: Application.get_env(:prima_auth0_ex, :token_cache, EncryptedRedisTokenCache)
+  defp token_cache, do: Config.token_cache(EncryptedRedisTokenCache)
 end