+++ lib/prima_auth0_ex/token_provider/token_cache.ex
@@ -5,6 +5,6 @@ defmodule PrimaAuth0Ex.TokenProvider.TokenCache do
 
   alias PrimaAuth0Ex.TokenProvider.TokenInfo
 
-  @callback set_token_for(String.t(), TokenInfo.t()) :: :ok | {:error, any()}
-  @callback get_token_for(String.t()) :: {:ok, TokenInfo.t() | nil} | {:error, any()}
+  @callback set_token_for(atom(), String.t(), TokenInfo.t()) :: :ok | {:error, any()}
+  @callback get_token_for(atom(), String.t()) :: {:ok, TokenInfo.t() | nil} | {:error, any()}
 end