+++ test/prima_auth0_ex/token_provider/auth0_authorization_service_test.exs
@@ -8,13 +8,14 @@ defmodule PrimaAuth0Ex.TokenProvider.Auth0AuthorizationServiceTest do
 
   @invalid_auth0_response ~s<{"error": "I am an invalid response from auth0"}>
 
+  @test_audience "test"
+
   @sample_credentials %Auth0Credentials{
+    client: :default_client,
     base_url: "http://localhost",
     client_id: "client_id",
     client_secret: "client_secret"
   }
-
-  @test_audience "test"
 
   setup do
     bypass = Bypass.open()