+++ test/prima_auth0_ex/token_provider/auth0_jwks_kids_fetcher_test.exs
@@ -6,6 +6,7 @@ defmodule PrimaAuth0Ex.TokenProvider.AuthJwksKidsFetcherTest do
 
   @auth0_jwks_api_path "/.well-known/jwks.json"
   @sample_credentials %Auth0Credentials{
+    client: :default_client,
     base_url: "http://localhost",
     client_id: "client_id",
     client_secret: "client_secret"