+++ CHANGELOG.md
@@ -7,6 +7,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `:clients` config, which can now be used to configure multiple clients
+- `config :prima_auth0_ex, :clients, default_client: [...]` can be used to configure a default client, so that `token_for` and other methods can be used without specifying a client
+- `:redis` configuration, which is now separate from the clients' ones
+
+### Changed
+
+- `:clients` and `:server` now have an `:auth0_base_url` parameter
+- `token_for` and `refresh_token_for` have an additional `client` parameter, which defaults to `:default_client`
+
+### Removed
+
+- `config :prima_auth0_ex, :client` no longer works, `:clients, default_client: [...]` can be used instead
+- `config :prima_auth0_ex, auth0_base_url: ...` is not used anymore, each base url is now client and server specific
+
 ---
 
 ## [0.5.0] - 2022-09-28
@@ -200,7 +216,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fixed compilation error when `:auth0_ex, :server` is not configured in `config.exs`
-
 
 [Unreleased]: https://github.com/primait/auth0_ex/compare/0.5.0...HEAD
 [0.5.0]: https://github.com/primait/auth0_ex/compare/0.4.6...0.5.0